requirements.txt

python-dotenv
openai
requests
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
chromadb
faiss-cpu
pypdf
Pillow
PyMuPDF
pdf2image